Redesigned home screen
Subject are now pinnable to the home page.

The home page shows pin and subject list.
Pin list has extra information and subject list is an openable list.

Fix result calculation that computed NaN when there's a question with
zero correct answers.

Set preload css to false, this was on before for setting background
color faster. The index.css was added to load the default theme first
instead.