Rename and refactor Collection to Deck
Add useContext with MonkeContext. This make sharing the global data
between the components easier.

Rewrote the MonkeCollection to DeckDB. The way it's interface is more
like database query. It stores the Deck information such as title and
its card list as ID form. The deck can be query by looking up its ID,
which is generated by hashing the source URL.

The cards are now stored separately into its own database and can be
look up  with its source hash as the key. The CardDB stores all the
cards the app will use.

Update katex css version.
Minor changes to the logo and build scripts.